Rename FeedCampaign to CampaignPost and delete FeedUpdate
